Refresh mirror list from an online location
Refresh the available mirror list from the list that is available in one
of the salix mirrors, in the MIRRORS file. First try to refresh the list
from the current repository and if that fails try every other repo in
the existing list until one works.

Also add the -l switch that disables the downloading of the new mirror
list and just uses the already present local list.