Load locales with respect to $T_PX
Add code that checks if we're running on an installer image, or we are
in an installed system and use that to set $T_PX. Once that is done,
load locales from $T_PX/usr/share/locale. Otherwise netsetup is not
translated when launched from the installer using setup.netsetup.