Update netsetup to catch up with netconfig in 15.0
Lot's of changes here, including IPv6 support. Translations need to
catch up, as many strings have been changed. I tried to keep the changes
to a minimum, but it wasn't always possible. In cases where there was to
be a change in a long string, I took the opportunity of splitting the
long string to multiple smaller ones. Let's hope I didn't miss a quote
or a backtick somewhere (fingers crossed)...